Unbreak secp256r1 on big endian by disabling the optimized path (#9)
Neale seems to have tried to fix this for s390x, but it's still not
working there nor on ppc64. We can just fall back to a less
optimized implementation however, which does work fine.

Upstream OpenSSL, LibreSSL, and BoringSSL have diverged heavily
from this implementation anyways; so we'd need to rework support
anyways.