[ANDROID] Fix another Android toolchain issue with crypto assembly
Some armv8 assembly sources contain the following directive:

    #if !defined(__clang__)
       .arch  armv8-a+crypto
     #endif

Which supposedly works using desktop (probably Apple) versions of Clang but, for
some reason, fails to build with Android NDK's clang - it appears the directive
is necessary for that compiler to properly compile the assembly source.

Make sure we do not exclude the directive when building for Android.